Fix discovery of program number by articulation
When the JSFX was reloaded (offlined and onlined), programs_by_articulation
would be incorrectly regenerated because program,channel tuples without any
assigned articulation would end up clobbering the mapping for whatever the
first articulation in the bank was.  This is because there was no way
to distinguish between unassigned and the first articulation: both had an
offset value of 0 in the articulations_offset array.

So the solution is to initialize articulations_offset to -1 rather than 0.