Improve box layout logic
Boxes now support alignment on the opposite dimension (e.g.
horizontal alignment for vertical boxes) of expanded children.

Fill attribute for opposite dimension (e.g. fillw for vertical
boxes) now supports FILL_TO_SIBLINGS which will cause the widget
to expand in that direction to be the same size as the largesst
sibling (with that dimension).