Revamp mainapp/RFX intercommunication and fix several bugs
This is a significant design overhaul of the way the JSFX and main
script communicate with one another.  Rather than using sliders
to move data in and out, we switch to the gmem buffer instead.
Consequently this now requires Reaper 5.97 which introduced the
Lua APIs for gmem access.

Closes #62 primarily, but in the process also fixes #44 and closes #47.
Also lays the groundwork for implementing CCs in the GUI described
in #41.