Ensure PC names in arrange view are resolved
Leverage new actions introduced in REAPER 6.46 to allow force-reloading
track support data (which includes Reabanks) when projects are opened,
or when banks are changed in track configuration.

This change also makes some significant effort to *avoid* calling these
actions, as they're quite expensive.  A persisted cache is maintained
that remembers what banks were previously written to the tmp Reabank,
and only invokes the new actions if there were material changes made.

Moreover, the reabank module now tries to avoid rewriting the rebank in
cases where it makes no difference to REAPER: when creating a new empty
project (which includes closing and reopening the current project), or
when removing banks from the project.