Added suggested utility files
###SUMMARY

Since Reaticulate is not in ReaPack, there is currently no way to easily synchronize new or modified .reabank files. This script is a Windows PowerShell script designed to help users sync their .reabank files to the latest Github userbanks.

### CONTENTS

`merge-userbanks-tree.ps1`

This script should be run from the `\reaticulate\userbanks\` folder. This obviously presupposes that you have downloaded the latest version of the Reaticulate master branch on Github. It will do the following:

- Merge all .reabank folders under the folder you run it in (which should be the `\reaticulate\userbanks\` folder downloaded from Github), noting the source folder and file of each.
- Rename your current Reaticulate.reabank file to Reaticulate.reabank.BAK. If the backup file already exists, the new one will have an incremental number appended (e.g. .BAK01).
- Save the new Reaticulate.reabank file to Reaper's Data folder (This assumes a standard Reaper installation, not a portable one).
- Normalize all line endings to CR/LF and eliminate extra newlines, except for before and after the source file indicator.

### REQUEST

I would request that this file be added inside `\reaticulate\userbanks\` and a mention of its purpose be added to the page's readme file.